Almost done
last things to add are polish and an inintial level in which the player is introduced to the concept of the game, a sort of tutorial level. Else to do is to create own sprites and particles for a few things. Adding sound is also a big todo. But it's good for now